Fixed micrometerMetricsRecorder
